+++ RELEASES.md
@@ -2,6 +2,7 @@
 
 * fixed Ec2SecurityGroups: error when egress/ingress null
 * RespondsToPing: Hostname not used
+* added dependency for 'aws-ssh-resolver'
 
 ## 0.2.3/20150921-14:08:51